Update App.js
Update holdDice function to setDice from state by looking through old dice, getting the id to see if it matches the id that was passed to the function. If it matches, then, the properties of the die should be spread, and the isHeld property is going to be set as the contrary to what it had previously.
If the ids don't match, then just return the die.