Update App.js
Import useEffect and use State
Import Confetti
Add state for Tenzies default as false
useEffect to create variables:
First allHeld, to identify if every die has isHeld prop set to true
Then firstValue to set value of the dice to the value of the first index array
Then allSameValue to see if every dice has their value equal to the firstValue variable value
And if all dice are held true and have the same value as the first one of the array, then we want to set Tenzies to true, console.log you won.

Now, if it's not tenzies, then we should setDice looping through every die and return die if the isHeld is true, or generate a new value if the isHeld is false. If it's tenzies, then we should set tenzies to false and then setDice to generate allNewDice.

At the main, we have a Confetti component that loads if we have tenzies

And, if we have tenzies, then the button turns its text into New Game, if not, then it turns into Roll.